ランキング、パーツ表示を作成

+++ HBHStump/Assets/Scripts/AnotherGameMode/MasterData.cs
@@ -13,7 +13,7 @@ public class MasterData : MonoBehaviour
     //ゲームのステータス
     public float remainingTime = 60*3;    //此方の変数を減らしていく
     public float remainingTimeTemp;       //制限時間は何秒か記憶用
-    public float score;            //得点
+    public int score;            //得点
 
     //変化させたオブジェクト
     public List<Sprite> newObjects = new List<Sprite>();
